Move Dto related to authentication and user controller to TimeScout.Application project
