+++ TimeScout.Application/DTOs/Authentication/LoginResponseDto.cs
@@ -1,10 +1,7 @@
-using System;
-
-namespace TimeScout.API.DTOs.Login;
+namespace TimeScout.Application.DTOs.Login;
 
 public class LoginResponseDto
 {
     public string AccessToken { get; set; } = null!;
-    public string RefreshToken { get; set; } = null!;
     public int UserId { get; set; }
 }
