+++ TimeScout.Application/DTOs/Authentication/RegisterRequestDto.cs
@@ -1,0 +1,10 @@
+namespace TimeScout.Application.DTOs.Authentication;
+
+public class RegisterRequestDto
+{
+    public string Username { get; set; } = null!;
+    public string Email { get; set; } = null!;
+    public string Password { get; set; } = null!;
+    public string FirstName { get; set; } = null!;
+    public string LastName { get; set; } = null!;
+}
