+++ TimeScout.Application/DTOs/User/UserUpdateRequestDto.cs
@@ -1,12 +1,11 @@
 using System;
 
-namespace TimeScout.API.DTOs.Authentication;
+namespace TimeScout.Application.DTOs.User;
 
-public class RegisterRequestDto
+public class UserUpdateRequestDto
 {
+    public int Id { get; set; }
     public string Username { get; set; } = null!;
-    public string Email { get; set; } = null!;
-    public string Password { get; set; } = null!;
     public string FirstName { get; set; } = null!;
     public string LastName { get; set; } = null!;
 }
